// src/api/__generated__/recommend/recommend-feed/recommend-feed.ts
export type feedResponse200 = {
  data: ResultFeedResponse;
  status: 200;
};

export type feedResponse400 = {
  data: ResultObject;
  status: 400;
};

export type feedResponseSuccess = feedResponse200 & {
  headers: Headers;
};
export type feedResponseError = feedResponse400 & {
  headers: Headers;
};

export type feedResponse = feedResponseSuccess | feedResponseError;